[Events] Updates tmea round table with cms data
