+++ src/pages/events/events-links.yml
@@ -9,6 +9,6 @@
     - title: Fall Golf Tournament
       link: /events/fall-golf-tournament/
     - title: TMEA Round Table
-      link: /events/winter-round-table/
+      link: /events/tmea-round-table/
     - title: Job Fair / Job Fair Rules
       link: /events/job-fair/
